// src/app/model/errors/analise-pre-compilacao/enum/tiposErrosRepeticoes.ts
export enum TiposErrosRepeticoes {
  faltaDoisPontos = "Falta dois pontos",
  semComparacao = "Repetição sem comparação",
  andOrMaiusculo = "Operador lógico maiúsculo",
  apenasUmParComparacao = "Apenas um dado na comparação",
  faltaDoisPontosCondicao = "Falta dois pontos na repetição",
  comparacaoApenasUmaIgualdade = "Apenas uma igualdade na repetição",
  faltaOperadorIn = "Faltou operador in",
  semParDadosIn = "Faltou um par de dados para o operador in",
  sintaxeWhileInvalida = "Sintaxe do while sem comparação",
  faltaParentesis = "Faltou um parêntesis de abertura ou fechamento"
}